# specs/002-intelligent-message-analysis/contracts/api-contract.yaml
openapi: 3.0.3
info:
  title: Intelligent Message Analysis API
  description: API endpoints for AI-powered message analysis, spam detection, and reply generation
  version: 1.0.0
  contact:
    name: MyAssistantPersonal Team
servers:
  - url: http://localhost:3002/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /messages/{messageId}/analysis:
    get:
      summary: Get message analysis
      description: Retrieve AI analysis for a specific message
      operationId: getMessageAnalysis
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          description: ID of the message to get analysis for
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Analysis retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageAnalysisResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Trigger message analysis
      description: Trigger AI analysis for a specific message (creates MessageAnalysis if not exists)
      operationId: analyzeMessage
      tags:
        - Messages
      parameters:
        - name: messageId
          in: path
          required: true
          description: ID of the message to analyze
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '201':
          description: Analysis created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageAnalysisResponse'
        '200':
          description: Analysis already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageAnalysisResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '503':
          description: Ollama AI service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Service Unavailable
                message: Ollama AI is currently offline
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageAnalysisResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/MessageAnalysis'
    MessageAnalysis:
      type: object
      required:
        - messageId
        - analysisTimestamp
        - analysisVersion
        - spamProbability
        - isSpam
        - needsResponse
        - responseConfidence
        - sentiment
        - priorityLevel
      properties:
        messageId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          example: '507f1f77bcf86cd799439011'
        analysisTimestamp:
          type: string
          format: date-time
          example: '2025-11-19T10:30:00Z'
        analysisVersion:
          type: string
          example: '1.0.0'
        spamProbability:
          type: integer
          minimum: 0
          maximum: 100
          example: 95
        isSpam:
          type: boolean
          example: true
        needsResponse:
          type: boolean
          example: false
        responseConfidence:
          type: integer
          minimum: 0
          maximum: 100
          example: 20
        sentiment:
          type: string
          enum: [positive, neutral, negative]
          example: 'negative'
        priorityLevel:
          type: string
          enum: [high, medium, low]
          example: 'low'
        generatedReplyText:
          type: string
          example: 'Thank you for your message. I will get back to you shortly.'
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Bad Request'
        message:
          type: string
          example: 'Invalid message ID format'
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
            message: Authentication token required
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not Found
            message: Message not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Internal Server Error
            message: An unexpected error occurred